requirements.txt
# Path: requirements.txt

requests==2.31.0
PyYAML==6.0.1
pydantic==2.7.1
gql[requests]~=3.4.0
googletrans==4.0.0-rc1
backoff~=2.2.1
schedule~=1.2.0